+++ docs/whats-new.rst
@@ -23,7 +23,7 @@ v0.1.14 (XX Xxx. 2023)
         f = CTDRefDataFetcher(box=[15, 30, -70, -60, 0, 5000.0])
         ds = f.to_xarray()
 
-- Index store can now export search results to standard Argo index files
+- Index store can now export search results to standard Argo index files. (pr:`260`) by `G. Maze <http://www.github.com/gmaze>`_
 
 **Internals**